21/09 18:00

+++ ProjetoAPI/Controllers/ProdutosController.cs
@@ -8,6 +8,7 @@ using System.Net;
 using System.Net.Http;
 using System.Web.Http;
 using System.Web.Http.Description;
+using ProjetoAPI.Models;
 using ProjetoAPI.Models.Context;
 using ProjetoAPI.Models.Entities;
 
@@ -17,15 +18,17 @@ namespace ProjetoAPI.Controllers
     {
         private BancoContext db = new BancoContext();
 
-        // GET: api/Produtos
+        Usuario user = new Usuario();
+
+        [Authorize(Roles = "Admin, User")]
         public IQueryable<Produto> GetProdutos()
         {
             return db.Produtos;
         }
 
-        // GET: api/Produtos/5
+        [Authorize(Roles = "Admin, User")]
         [ResponseType(typeof(Produto))]
-        public IHttpActionResult GetProduto(int id)
+        public IHttpActionResult GetProdutoId(int id)
         {
             Produto produto = db.Produtos.Find(id);
             if (produto == null)
@@ -36,10 +39,28 @@ namespace ProjetoAPI.Controllers
             return Ok(produto);
         }
 
-        // PUT: api/Produtos/5
+        ////[Authorize(Roles = "Admin")]
+        //[HttpGet, Route("api/produtos/{nome?}")]
+        //[ResponseType(typeof(Produto))]
+        //public IHttpActionResult GetProdutoNome(string nome)
+        //{
+        //    Produto produto = db.Produtos.SingleOrDefault(m => m.Nome == nome);
+
+
+        //    if (produto == null)
+        //    {
+        //        return NotFound();
+        //    }
+
+        //    return Ok(produto);
+        //}
+
+        [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "Admin")]
         [ResponseType(typeof(void))]
         public IHttpActionResult PutProduto(int id, Produto produto)
         {
+
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -50,11 +71,20 @@ namespace ProjetoAPI.Controllers
                 return BadRequest();
             }
 
+
             db.Entry(produto).State = EntityState.Modified;
+            produto.DataAlteracao = DateTime.Now;
 
             try
             {
-                db.SaveChanges();
+                if (user.Ativo == true)
+                {
+                    db.SaveChanges();
+                }
+                else
+                {
+                    return Json("Erro, usuario inativo.");
+                }
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -71,22 +101,37 @@ namespace ProjetoAPI.Controllers
             return StatusCode(HttpStatusCode.NoContent);
         }
 
-        // POST: api/Produtos
+        [Authorize(Roles = "Admin")]
         [ResponseType(typeof(Produto))]
         public IHttpActionResult PostProduto(Produto produto)
         {
+
+            Usuario usuario = new Usuario();
+
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
 
-            db.Produtos.Add(produto);
-            db.SaveChanges();
+            produto.DataCadastro = DateTime.Now;
 
-            return CreatedAtRoute("DefaultApi", new { id = produto.Id }, produto);
+            var teste = usuario.GetEnable();
+
+            if (user.Ativo == true)
+            {
+                db.Produtos.Add(produto);
+                db.SaveChanges();
+                return CreatedAtRoute("DefaultApi", new { id = produto.Id }, produto);
+            }
+            else
+            {
+                return Json("Erro, usuario inativo.");
+
+            }
+
         }
 
-        // DELETE: api/Produtos/5
+        [Authorize(Roles = "Admin")]
         [ResponseType(typeof(Produto))]
         public IHttpActionResult DeleteProduto(int id)
         {
@@ -96,10 +141,17 @@ namespace ProjetoAPI.Controllers
                 return NotFound();
             }
 
-            db.Produtos.Remove(produto);
-            db.SaveChanges();
+            if (user.Ativo)
+            {
+                db.Produtos.Remove(produto);
+                db.SaveChanges();
 
-            return Ok(produto);
+                return Ok(produto);
+            }
+            else
+            {
+                return Json("Erro, usuario inativo.");
+            }
         }
 
         protected override void Dispose(bool disposing)
